addBBands {quantmod}	R Documentation

Add Bollinger Bands to Chart

Description

Add Bollinger Bands to current chart.

Usage

addBBands(n = 20, sd = 2, maType = "SMA", draw = 'bands', on = -1)

Arguments

n
number of moving average periods

maType
type of moving average to be used

sd
number of standard deviations

draw
indicator to draw: bands, percent, or width

on
which figure area of chart to apply to

Details

The primary addition to this function call over the TTR version is in the draw argument. ‘bands’ will draw standard Bollinger Bands, ‘percent’ will draw Bollinger %b and ‘width’ will draw Bolinger Bands Width. The last two will be drawn in new figure regions.

See bollingerBands in TTR for specific details as to implementation and references.

Value

Bollinger Bands will be drawn, or scheduled to be drawn, on the current chart. If draw is either percent or width a new figure will be added to the current TA figures charted.

A chobTA object will be returned silently.

Author(s)

Jeffrey A. Ryan

References

See bollingerBands in TTR written by Josh Ulrich

See Also

addTA

Examples

## Not run: 
addBBands()

## End(Not run)

[Package quantmod version 0.4.18 Index]