addSAR {quantmod}	R Documentation

Add Parabolic Stop and Reversal to Chart

Description

Add Parabolic Stop and Reversal indicator overlay to chart.

Usage

addSAR(accel = c(0.02, 0.2), col = "blue")

Arguments

accel
Accelleration factors - see SAR

col
color of points (optional)

Details

see 'SAR' in TTR for specific details and references.

Value

A SAR overlay will be drawn on the current chart. A chobTA object will be returned silently.

Author(s)

Jeffrey A. Ryan

References

see SAR in TTR written by Josh Ulrich

See Also

addTA

Examples

## Not run: 
addSAR()

## End(Not run)

[Package quantmod version 0.4.18 Index]