addVo {quantmod}	R Documentation

Add Volume to Chart

Description

Add Volume of a series, if available, to the current chart. This is the default TA argument for all charting functions.

Usage

addVo(log.scale=FALSE)

Arguments

log.scale
use log-scale for volume

Details

Add volume bars to current chart if data object contains appropriate volume column.

log.scale will transform the series via standard R graphics mechanisms.

Value

Volume will be draw in a new window on the current chart. A chobTA object will be returned silently.

Author(s)

Jeffrey A. Ryan

See Also

addTA

Examples

## Not run: 
addVo()

## End(Not run)

[Package quantmod version 0.4.18 Index]