chart_Series {quantmod}	R Documentation

Experimental Charting Version 2

Description

These are experimental functions for a new version of chartSeries in quantmod. Interface, behavior, and functionality will change.

Usage

chart_Series(x,
             name = deparse(substitute(x)), 
             type = "candlesticks", 
             subset = "",
             TA = "", 
             pars = chart_pars(), 
             theme = chart_theme(), 
             clev = 0,
             ...)

Arguments

x
time series object

name
name for chart

type
one of:

subset
an ISO8601 style character string indicating date range

TA
a character string of semi-colon seperated TA calls.

pars
chart parameters

theme
chart theme

clev
color level (experimental). Indicates the degree of brightness 0 is darkest color.

...
additional parameters

Details

These functions, when complete, will revert back to the original chartSeries naming convention.

Value

Called for graphical side effects.

Note

Highly experimental (read: alpha) use with caution.

Author(s)

Jeffrey A. Ryan

[Package quantmod version 0.4.18 Index]