chobTA-class {quantmod}	R Documentation

A Technical Analysis Chart Object

Description

Internal storage class for handling arbitrary TA objects

Objects from the Class

Objects of class chobTA are created and returned invisibly through calls to addTA-style functions.

Slots

call:

Object of class "call" ~~

on:

Object of class "ANY" ~~

new:

Object of class "logical" ~~

TA.values:

Object of class "ANY" ~~

name:

Object of class "character" ~~

params:

Object of class "ANY" ~~

Methods

show

signature(object = "chobTA"): ...

Note

It is of no external vaule to handle chobTA objects directly

Author(s)

Jeffrey A. Ryan

See Also

addTA, ~~~ or chob for links to other classes

Examples

showClass("chobTA")

[Package quantmod version 0.4.18 Index]