create.binding {quantmod}	R Documentation

Create DDB Bindings

Description

Internal function used in attachSymbols to create active bindings for symbols defined in a DDB object.

Usage

create.binding(s,
               lsym,
               rsym,
               gsrc,
               mem.cache = TRUE,
               file.cache = !mem.cache,
               cache.dir = tempdir(),
               envir,...)

Arguments

s
symbol name

lsym
function to convert to local name (legal R name)

rsym
function to convert to remote name (source name)

gsrc
corresponds to 'src' arg in getSymbols call

mem.cache
cache to memory

file.cache
cache to disk

cache.dir
directory to cache to/from

envir
environment name (character)

...
arguments to pass to getSymbols call

Details

Low level function to create bindings during initial demand-database construction.

Value

Called for its side effect of creating active bindings to symbols.

Note

This is code used internally by attachSymbols. User's may modify this to accomodate different systems. The upstream functions needn't maintain consistency with this interface.

Use as a guide or template.

Author(s)

Jeffrey A. Ryan

References

Mark, Roger, ?

[Package quantmod version 0.4.18 Index]