fittedModel {quantmod}	R Documentation

quantmod Fitted Objects

Description

Extract and replace fitted models from quantmod objects built with buildModel. All objects fitted through methods specified in buildModel calls can be extracted for further analysis.

Usage

fittedModel(object)

## S3 method for class 'quantmod'
formula(x, ...)

## S3 method for class 'quantmod'
plot(x, ...)

## S3 method for class 'quantmod'
coefficients(object, ...)

## S3 method for class 'quantmod'
coef(object, ...)

## S3 method for class 'quantmod'
residuals(object, ...)

## S3 method for class 'quantmod'
resid(object, ...)

## S3 method for class 'quantmod'
fitted.values(object, ...)

## S3 method for class 'quantmod'
fitted(object, ...)

## S3 method for class 'quantmod'
anova(object, ...)

## S3 method for class 'quantmod'
logLik(object, ...)

## S3 method for class 'quantmod'
vcov(object, ...)

Arguments

object
a quantmod object

x
a suitable object

...
additional arguments

Details

Most often used to extract the final fitted object of the modelling process, usually for further analysis with tools outside the quantmod package.

Most common methods to apply to fitted objects are available to the parent quantmod object. At present, one can call directly the following S3 methods on a built model as if calling directly on the fitted object. See *Usage* section.

It is also possible to add a fitted model to an object. This may be of value when applying heuristic rule sets for trading approaches, or when fine tuning already fit models by hand.

Value

Returns an object matching that returned by a call to the method specified in buildModel.

Note

The replacement function fittedModel<- is highly experimental, and may or may not continue into further releases.

The fitted model added must use the same names as appear in the quantmod object's dataset.

Author(s)

Jeffrey A. Ryan

See Also

quantmod,buildModel

Examples

## Not run: 
x <- specifyModel(Next(OpCl(DIA)) ~ OpCl(VIX))
x.lm <- buildModel(x,method='lm',training.per=c('2001-01-01','2001-04-01'))

fittedModel(x.lm)

coef(fittedModel(x.lm))
coef(x.lm)                  # same

vcov(fittedModel(x.lm))
vcov(x.lm)                  # same

## End(Not run)

[Package quantmod version 0.4.18 Index]