getFX {quantmod}	R Documentation

Download Exchange Rates

Description

Download exchange rates or metals prices from oanda.

Usage

getFX(Currencies,
      from = Sys.Date() - 179,
      to = Sys.Date(),
      env = parent.frame(),
      verbose = FALSE,
      warning = TRUE,
      auto.assign = TRUE, ...)

Arguments

Currencies
Currency pairs expressed as ‘CUR/CUR’

from
start date expressed in ISO CCYY-MM-DD format

to
end date expressed in ISO CCYY-MM-DD format

env
which environment should they be loaded into

verbose
be verbose

warning
show warnings

auto.assign
use auto.assign

...
additional parameters to be passed to getSymbols.oanda method

Details

A convenience wrapper to getSymbols(x,src='oanda'). See getSymbols and getSymbls.oanda for more detail.

Value

The results of the call will be the data will be assigned automatically to the environment specified (parent by default). Additionally a vector of downloaded symbol names will be returned.

See getSymbols and getSymbols.oanda for more detail.

Author(s)

Jeffrey A. Ryan

References

Oanda.com https://www.oanda.com

See Also

getSymbols, getSymbols.oanda

Examples

## Not run: 

getFX("USD/JPY")

getFX("EUR/USD",from="2005-01-01")

## End(Not run)

[Package quantmod version 0.4.18 Index]