getFinancials {quantmod}	R Documentation

Download and View Financial Statements

Description

Download Income Statement, Balance Sheet, and Cash Flow Statements.

Usage

getFinancials(Symbol, env = parent.frame(), src = "google", 
              auto.assign = TRUE,
              ...)

viewFinancials(x, type=c('BS','IS','CF'), period=c('A','Q'),
               subset = NULL)

Arguments

Symbol
one or more valid google symbol, as a character vector or semi-colon delimited string

env
where to create the object

src
currently no sources are implemented

auto.assign
should results be loaded to the environment

...
currently unused

x
an object of class financials

type
type of statement to view

period
period of statement to view

subset
‘xts’ style subset string

Details

A utility to download financial statements for publicly traded companies.

Individual statements can be accessed using standard R list extraction tools, or by using viewFinancials.

viewFinancials allows for the use of date subsetting as available in the xts package, as well as the specification of the type of statement to view. BS for balance sheet, IS for income statement, and CF for cash flow statement. The period argument is used to identify which statements to view - (A) for annual and (Q) for quarterly.

Value

Six individual matrices organized in a list of class ‘financials’:

IS
a list containing (Q)uarterly and (A)nnual Income Statements

BS
a list containing (Q)uarterly and (A)nnual Balance Sheets

CF
a list containing (Q)uarterly and (A)nnual Cash Flow Statements

Note

As with all free data, you may be getting exactly what you pay for. Sometimes that may be absolutely nothing.

Author(s)

Jeffrey A. Ryan

Examples

## Not run: 
getFinancials('JAVA')  # returns JAVA.f to "env"
getFin('AAPL') # returns AAPL.f to "env"

viewFin(JAVA.f, "IS", "Q")  # Quarterly Income Statement
viewFin(AAPL.f, "CF", "A")   # Annual Cash Flows

str(AAPL.f)

## End(Not run)

[Package quantmod version 0.4.18 Index]