getMetals {quantmod}	R Documentation

Download Daily Metals Prices

Description

Download daily metals prices from oanda.

Usage

getMetals(Metals,
      from = Sys.Date() - 179,
      to = Sys.Date(),
      base.currency="USD",
      env = parent.frame(),
      verbose = FALSE,
      warning = TRUE,
      auto.assign = TRUE, ...)

Arguments

Metals
metals expressed in common name or symbol form

from
start date expressed in ISO CCYY-MM-DD format

to
end date expressed in ISO CCYY-MM-DD format

base.currency
which currency should the price be in

env
which environment should they be loaded into

verbose
be verbose

warning
show warnings

auto.assign
use auto.assign

...
additional parameters to be passed to getSymbols.oanda method

Details

A convenience wrapper to getSymbols(x,src='oanda').

The most useful aspect of getMetals is the ablity to specify the Metals in terms of underlying 3 character symbol or by name (e.g. XAU (gold) , XAG (silver), XPD (palladium), or XPT (platinum)).

There are unique aspects of any continuously traded commodity, and it is recommended that the user visit https://www.oanda.com for details on specific pricing issues.

See getSymbols and getSymbls.oanda for more detail.

Value

Data will be assigned automatically to the environment specified (parent by default). If auto.assign is set to FALSE, the data from a single metal request will simply be returned from the function call.

If auto.assign is used (the default) a vector of downloaded symbol names will be returned.

See getSymbols and getSymbols.oanda for more detail.

Author(s)

Jeffrey A. Ryan

References

Oanda.com https://www.oanda.com

See Also

getSymbols, getSymbols.oanda

Examples

## Not run: 

getFX(c("gold","XPD"))

getFX("plat",from="2005-01-01")

## End(Not run)

[Package quantmod version 0.4.18 Index]