getModelData {quantmod}	R Documentation

Update model's dataset

Description

Update currently specified or built model with most recent data.

Usage

getModelData(x, na.rm = TRUE)

Arguments

x
An object of class quantmod

na.rm
Boolean. Remove NA values. Defaults to TRUE

Details

Primarily used within specify model calls, getModelData is used to retrieve the appropriate underlying variables, and apply model specified transformations automatically. It can be used to also update a current model in memory with the most recent data.

Value

Returns object of class quantmod.OHLC

Author(s)

Jeffrey Ryan

See Also

getSymbols load data specifyModel create model structure buildModel construct model modelData extract model dataset

Examples

## Not run: 
my.model <- specifyModel(Next(OpCl(QQQQ)) ~ Lag(Cl(NDX),0:5))
getModelData(my.model)

## End(Not run)

[Package quantmod version 0.4.18 Index]