getSymbols.SQLite {quantmod}	R Documentation

Retrieve Data from SQLite Database

Description

Fetch data from SQLite database. As with other methods extending getSymbols this function should NOT be called directly.

Usage

getSymbols.SQLite(Symbols,
                  env,
                  return.class = 'xts',
                  db.fields = c("row_names",
                                "Open",
                                "High",
                                "Low",
                                "Close",
                                "Volume",
                                "Adjusted"),
                  field.names = NULL,
                  dbname = NULL,
                  POSIX = TRUE,
                  ...)

Arguments

Symbols
a character vector specifying the names of each symbol to be loaded

env
where to create the objects

return.class
desired class of returned object

db.fields
character vector naming fields to retrieve

field.names
names to assign to returned columns

dbname
database name

POSIX
are rownames numeric

...
additional arguments

Details

Meant to be called internally by getSymbols (see also)

One of a few currently defined methods for loading data for use with 'quantmod'. Its use requires the packages 'DBI' and 'RSQLite', along with a SQLite database.

The purpose of this abstraction is to make transparent the 'source' of the data, allowing instead the user to concentrate on the data itself.

Value

A call to getSymbols.SQLite will load into the specified environment one object for each 'Symbol' specified, with class defined by 'return.class'.

Note

This function is experimental at best, and has not been thoroughly tested. Use with caution, and please report any bugs to the maintainer of quantmod.

Author(s)

Jeffrey A. Ryan

References

SQLite https://www.sqlite.org/index.html

David A. James RSQLite: SQLite interface for R

R-SIG-DB. DBI: R Database Interface

See Also

getSymbols

Examples

## Not run: 
getSymbols("QQQQ",src="SQLite")

## End(Not run)

[Package quantmod version 0.4.18 Index]