getSymbols.oanda {quantmod}	R Documentation

Download Currency and Metals Data from Oanda.com

Description

Access to 191 currency and metal prices, downloadable as more that 36000 currency pairs from Oanda.com.

Downloads Symbols to specified env from https://www.oanda.com historical currency database. This method is not meant to be called directly, instead a call to getSymbols("x",src="oanda") will in turn call this method. It is documented for the sole purpose of highlighting the arguments accepted, and to serve as a guide to creating additional getSymbols 'methods'.

Usage

getSymbols.oanda(Symbols,
                 env,
                 return.class = "xts",
                 from = Sys.Date() - 179,
                 to = Sys.Date(),
                 ...)

Arguments

Symbols
a character vector specifying the names of each symbol to be loaded - expressed as a currency pair. (e.g. U.S. Dollar to Euro rate would be expressed as a string “USD/EUR”. The naming convention follows from Oanda.com, and a table of possible values is available by calling oanda.currencies

env
where to create objects.

return.class
class of returned object

from
Start of series expressed as "CCYY-MM-DD"

to
Start of series expressed as "CCYY-MM-DD"

...
additional parameters

Details

Meant to be called internally by getSymbols only.

Oanda data is 7 day daily average price data, that is Monday through Sunday. Oanda only provides historical data for the past 180 days. getSymbols will return as much data as possible, and warn when the from date is more than 180 days ago.

Value

A call to getSymbols(Symbols,src="oanda") will load into the specified environment one object for each 'Symbol' specified, with class defined by 'return.class'. Presently this may be 'ts', 'zoo', 'xts', or 'timeSeries'.

Note

Oanda rates are quoted as one unit of base currency to the equivelant amount of foreign currency.

Author(s)

Jeffrey A. Ryan

References

Oanda.com https://www.oanda.com

See Also

Currencies: getSymbols.FRED, getSymbols

Examples

## Not run: 
getSymbols("USD/EUR",src="oanda")
getSymbols("USD/EUR",src="oanda",from="2005-01-01")

## End(Not run)

[Package quantmod version 0.4.18 Index]