modelData {quantmod}	R Documentation

Extract Dataset Created by specifyModel

Description

Extract from a quantmod object the dataset created for use in modelling.

specifyModel creates a zoo object for use in subsequent workflow stages ( buildModel,tradeModel) that combines all model inputs, from a variety of sources, into one model frame.

modelData returns this object.

Usage

modelData(x, data.window = NULL, exclude.training = FALSE)

Arguments

x
a quantmod object

data.window
a character vector of subset start and end dates to return

exclude.training
remove training period

Details

When a model is created by specifyModel, it is attached to the returned object. One of the slots of this S4 class is model.data.

Value

an object of class zoo containing all transformations to data specified in specifyModel.

Author(s)

Jeffrey A. Ryan

See Also

specifyModel,getModelData

Examples

## Not run: 
m <- specifyModel(Next(OpCl(SPY)) ~ Cl(SPY) + OpHi(SPY) + Lag(Cl(SPY)))
modelData(m) 

## End(Not run)

[Package quantmod version 0.4.18 Index]