modelSignal {quantmod}	R Documentation

Extract Model Signal Object

Description

Extract model signal object from quantmodResults object as an object of class zoo.

Usage

modelSignal(x)

Arguments

x
object of class quantmodResults

Details

For use after a call to tradeModel to extract the generated signal of a given quantmod model. Normally this would not need to be called by the end user unless he was manually post processing the trade results.

Value

A zoo object indexed by signal dates.

Author(s)

Jeffrey A. Ryan

See Also

tradeModel

[Package quantmod version 0.4.18 Index]