options.expiry {quantmod}	R Documentation

Calculate Contract Expirations

Description

Return the index of the contract expiration date. The third Friday of the month for options, the last third Friday of the quarter for futures.

Usage

options.expiry(x)
futures.expiry(x)

Arguments

x
a time-indexed zoo object

Details

Designed to be used within a charting context via addExpiry, the values returned are based on the description above. Exceptions, though rare, are not accounted for.

Value

A numeric vector of values to index on.

Note

There is currently no accounting for holidays that may interfere with the general rule. Additionally all efforts have been focused on US equity and futures markets.

Author(s)

Jeffrey A. Ryan

References

~put references to the literature/web site here ~

See Also

addExpiry

Examples

## Not run: 
getSymbols("AAPL")

options.expiry(AAPL)
futures.expiry(AAPL)

AAPL[options.expiry(AAPL)]

## End(Not run)

[Package quantmod version 0.4.18 Index]