quantmod-class {quantmod}	R Documentation

Class "quantmod"

Description

Objects of class quantmod help to manage the process of model building within the quantmod package. Created automatically by a call to specifyModel they carry information to be used by a variety of accessor functions and methods.

Objects from the Class

Objects can be created by calls of the form new("quantmod", ...).

Normally objects are created as a result of a call to specifyModel.

Slots

model.id:

Object of class "character" ~~

model.spec:

Object of class "formula" ~~

model.formula:

Object of class "formula" ~~

model.target:

Object of class "character" ~~

model.inputs:

Object of class "character" ~~

build.inputs:

Object of class "character" ~~

symbols:

Object of class "character" ~~

product:

Object of class "character" ~~

price.levels:

Object of class "zoo" ~~

training.data:

Object of class "Date" ~~

build.date:

Object of class "Date" ~~

fitted.model:

Object of class "ANY" ~~

model.data:

Object of class "zoo" ~~

quantmod.version:

Object of class "numeric" ~~

Methods

No methods defined with class "quantmod" in the signature.

Author(s)

Jeffrey A. Ryan

Examples

showClass("quantmod")

[Package quantmod version 0.4.18 Index]