quantmod-defunct {quantmod}	R Documentation

Defunct Functions in Package quantmod

Description

The functions listed here are no longer part of quantmod.

Usage

  # Defunct in 0.4-13
  getFin(Symbol, env = parent.frame(), src = "google", auto.assign = TRUE, ...)
  getFinancials(Symbol, env = parent.frame(), src = "google", auto.assign = TRUE, ...)
  getSymbols.google(Symbols, env, return.class = "xts", from = "2007-01-01",
    to = Sys.Date(), ...)

Details

Google Finance stopped providing data in March, 2018. This included historical price data, and also financial information. This affected getFinancials, getFin, and getSymbols.google.

See Also

Defunct

[Package quantmod version 0.4.18 Index]