quantmod-package {quantmod}	R Documentation

Quantitative Financial Modelling Framework

Description

Quantitative Financial Modelling and Trading Framework for R

Details

Package: 	 quantmod

Type: 	 Package

Title: 	 Quantitative Financial Modelling Framework

Version: 	 0.4.18

Authors@R: 	 c( person(given=c("Jeffrey","A."), family="Ryan", role=c("aut","cph")), person(given=c("Joshua","M."), family="Ulrich", role=c("cre","aut"), email="josh.m.ulrich@gmail.com"), person(given="Wouter", family="Thielen", role="ctb"), person(given="Paul", family="Teetor", role="ctb"), person(given="Steve", family="Bronder", role="ctb") )

Depends: 	 R (>= 3.2.0), xts(>= 0.9-0), zoo, TTR(>= 0.2), methods

Imports: 	 curl

Suggests: 	 DBI,RMySQL,RSQLite,timeSeries,xml2,downloader,jsonlite(>= 1.1)

Description: 	 Specify, build, trade, and analyse quantitative financial trading strategies.

LazyLoad: 	 yes

License: 	 GPL-3

URL: 	 http://www.quantmod.com https://github.com/joshuaulrich/quantmod

BugReports: 	 https://github.com/joshuaulrich/quantmod/issues

Author: 	 Jeffrey A. Ryan [aut, cph], Joshua M. Ulrich [cre, aut], Wouter Thielen [ctb], Paul Teetor [ctb], Steve Bronder [ctb]

Maintainer: 	 Joshua M. Ulrich <josh.m.ulrich@gmail.com>

The quantmod package for R is designed to assist the quantitative trader in the development, testing, and deployment of statistically based trading models.

What quantmod IS

A rapid prototyping environment, with comprehensive tools for data management and visualization. where quant traders can quickly and cleanly explore and build trading models.

What quantmod is NOT

A replacement for anything statistical. It has no 'new' modelling routines or analysis tool to speak of. It does now offer charting not currently available elsewhere in R, but most everything else is more of a wrapper to what you already know and love about the langauge and packages you currently use.

quantmod makes modelling easier by removing the repetitive workflow issues surrounding data management, modelling interfaces, and performance analysis.

Author(s)

Jeffrey A. Ryan

Maintainer: Joshua M. Ulrich <josh.m.ulrich@gmail.com>

[Package quantmod version 0.4.18 Index]