quantmod.OHLC {quantmod}	R Documentation

Create Open High Low Close Object

Description

Coerce an object with the apporpriate columns to class quantmod.OHLC, which extends zoo.

Usage

as.quantmod.OHLC(x,
                 col.names = c("Open", "High",
                               "Low", "Close",
                               "Volume", "Adjusted"),
                 name = NULL, ...)

Arguments

x
object of class zoo

col.names
suffix for columns

name
name to attach unique column suffixes to, defaults to the object name

...
additional arguments (unused)

Details

quantmod.OHLC is actually just a renaming of an object of class zoo, with the convention of NAME.Open, NAME.High, ... for the column names.

Additionally methods may be written to handle or check for the above conditions within other functions - as is the case within the quantmod package.

Value

An object of class c('quantmod.OHLC','zoo')

Author(s)

Jeffrey A. Ryan

See Also

OHLC.Transformations, getSymbols

[Package quantmod version 0.4.18 Index]