zoomChart {quantmod}	R Documentation

Change Zoom Level Of Current Chart

Description

Using xts style date subsetting, zoom into or out of the current chart.

Usage

zooom(n=1, eps=2)
zoomChart(subset, yrange=NULL)

Arguments

n
the number of interactive view changes per call

eps
the distance between clicks to be considered a valid subset request

subset
a valid subset string

yrange
override y-scale

Details

These function allow for viewing of specific areas of a chart produced by chartSeries by simply specifying the dates of interest

zooom is an interactive chart version of zoomChart which utilizes the standard R device interaction tool locator to estimate the subset desired. This estimate is then passed to zoomChart for actual redrawing. At present it is quite experimental in its interface and arguments. Its usage entails a call to zooom() followed by the selection of the leftmost and rightmost points desired in the newly zoomed chart. This selection is accomplished by the user left-clicking each extreme point. Two click are required to determine the level of zooming. Double clicking will reset the chart to the full data range. The arguments and internal working of this function are likely to change dramatically in future releases, though its use will likely remain.

Standard format for the subset argument is the same as the subsetting for xts objects, which is how the data is stored internally for rendering.

Calling zoomChart with no arguments (NULL) resets the chart to the original data.

Examples include '2007' for all of the year 2007, '2007::2008' for years 2007 through 2008, '::2007' for all data from the beginning of the set to the end of 2007, '2007::' all data from the beginning of 2007 through the end of the data. For specifics regarding the level of detail and internal interpretation please see [.xts

Value

This function is called for its side effect - notably changing the perspective of the current chart, and changing its formal subset level. The underlying data attached to the chart is left unchanged.

Author(s)

Jeffrey A. Ryan

See Also

chartSeries

Examples

## Not run: 
data(sample_matrix)
chartSeries(sample_matrix)
zoomChart('2007-04::')
zoomChart()

zooom() # interactive example

## End(Not run)

[Package quantmod version 0.4.18 Index]